[WFCORE-1618] Push ActiveOperationSupport down into its only subclass
